A2: Fix ntuple filters
use wider type for register value
set ipv4 configuration (6 Ipv6 + 8 IPv4)

+++ drivers/net/ethernet/aquantia/atlantic-fwd/atl_common.h
@@ -106,14 +106,14 @@ struct atl2_rxf_l3 {
 		};
 	};
 	u16 proto;
-	u16 cmd;
+	u32 cmd;
 	u16 usage;
 };
 
 struct atl2_rxf_l4 {
 	__be16 dst_port;
 	__be16 src_port;
-	u16 cmd;
+	u32 cmd;
 	u16 usage;
 };
